requirements.txt
APScheduler>=3.10.4
yt-dlp>=2024.1.1
youtube-transcript-api>=0.6.2
google-genai>=1.0.0
openai>=1.40.0
anthropic>=0.34.0
python-frontmatter>=1.1.0
PyYAML>=6.0.1
pytest>=8.0.0
pytest-mock>=3.14.0
